Render cursor in the minibuffer when needed

+++ src/dispatcher.rs
@@ -17,10 +17,11 @@ fn is_self_insert(keys: &Vec<Key>) -> Option<char> {
 /// Process user input.
 pub fn process_user_input(term: &mut Term, context: &mut Context) {
     let cmd = read::read_key_binding(term, context);
-
     let minibuffer = &mut context.buffer_list.minibuffer;
 
-    minibuffer.truncate();
+    if !context.buffer_list.minibuffer_focused {
+        minibuffer.truncate();
+    }
 
     // Execute the command.
     match cmd {
